others/abc075/c/src/main/java/com/solver/atcoder/others/abc075/c/Main.java
package com.solver.atcoder.others.abc075.c;

import java.io.InputStream;
import java.io.IOException;
import java.io.PrintStream;
import java.util.*;

public class Main {

  public static void main(String[] args) {
    solve(System.in, System.out);
  }

  private static class Pair {

    public Pair(int a, int b) {
      this.a = a;
      this.b = b;
    }

    int a;
    int b;
  }

  static void solve(InputStream is, PrintStream os) {
    Scanner sc = new Scanner(is);

    /* read */
    int n = sc.nextInt();
    int m = sc.nextInt();

    boolean[][] edges = new boolean[n][n];
    List<Pair> pairs = new ArrayList<>(m);

    for (int i = 0; i < m; i++) {
      int a = sc.nextInt() - 1;
      int b = sc.nextInt() - 1;
      edges[a][b] = true;
      edges[b][a] = true;
      pairs.add(new Pair(a, b));
    }

    int count = 0;
    for (Pair pair : pairs) {
      edges[pair.a][pair.b] = false;
      edges[pair.b][pair.a] = false;
      if (!isConnected(edges)) count++;
      edges[pair.a][pair.b] = true;
      edges[pair.b][pair.a] = true;
    }
    os.println(count);
  }

  private static boolean isConnected(boolean[][] edges) {
    Queue<Integer> queue = new ArrayDeque<>();
    queue.add(0);
    Set<Integer> visited = new HashSet<>();
    while (!queue.isEmpty()) {
      int node = queue.remove();
      visited.add(node);
      for (int to = 0; to < edges[node].length; to++) {
        if (edges[node][to] && !visited.contains(to)) queue.add(to);
      }
    }
    return edges.length == visited.size();
  }

  private static class Scanner {

    private final InputStream is;
    private final byte[] buffer = new byte[1024];
    private int ptr = 0;
    private int buflen = 0;

    Scanner(InputStream is) {
      this.is = is;
    }

    private boolean hasNextByte() {
      if (ptr < buflen) {
        return true;
      } else {
        ptr = 0;
        try {
          buflen = is.read(buffer);
        } catch (IOException e) {
          e.printStackTrace();
        }
        if (buflen <= 0) {
          return false;
        }
      }
      return true;
    }

    private int readByte() {
      if (hasNextByte()) {
        return buffer[ptr++];
      } else {
        return -1;
      }
    }

    private static boolean isPrintableChar(int c) {
      return 33 <= c && c <= 126;
    }

    public boolean hasNext() {
      while (hasNextByte() && !isPrintableChar(buffer[ptr])) {
        ptr++;
      }
      return hasNextByte();
    }

    public String next() {
      if (!hasNext()) {
        throw new NoSuchElementException();
      }
      StringBuilder sb = new StringBuilder();
      int b = readByte();
      while (isPrintableChar(b)) {
        sb.appendCodePoint(b);
        b = readByte();
      }
      return sb.toString();
    }

    public long nextLong() {
      if (!hasNext()) {
        throw new NoSuchElementException();
      }
      long n = 0;
      boolean minus = false;
      int b = readByte();
      if (b == '-') {
        minus = true;
        b = readByte();
      }
      if (b < '0' || '9' < b) {
        throw new NumberFormatException();
      }
      while (true) {
        if ('0' <= b && b <= '9') {
          n *= 10;
          n += b - '0';
        } else if (b == -1 || !isPrintableChar(b)) {
          return minus ? -n : n;
        } else {
          throw new NumberFormatException();
        }
        b = readByte();
      }
    }

    public int nextInt() {
      long nl = nextLong();
      if (nl < Integer.MIN_VALUE || nl > Integer.MAX_VALUE) {
        throw new NumberFormatException();
      }
      return (int) nl;
    }

    public double nextDouble() {
      return Double.parseDouble(next());
    }
  }
}
