others/abc005/c/src/test/java/com/solver/atcoder/others/abc005/c/MainTest.java
package com.solver.atcoder.others.abc005.c;

import java.io.ByteArrayInputStream;
import java.io.ByteArrayOutputStream;
import java.io.PrintStream;

import org.junit.Test;

import static org.hamcrest.CoreMatchers.is;
import static org.hamcrest.MatcherAssert.assertThat;

public class MainTest {

  @Test
  public void test() {
    String input = "1\n"
        + "3\n"
        + "1 2 3\n"
        + "3\n"
        + "2 3 4";
    String expected = "yes";
    
    assertThat(execute(input + "\n"), is(expected + "\n"));
  }

  @Test
  public void test1() {
    String input = "1\n"
        + "3\n"
        + "1 2 3\n"
        + "3\n"
        + "1 2 2";
    String expected = "no";

    assertThat(execute(input + "\n"), is(expected + "\n"));
  }

  private String execute(String input) {
    ByteArrayOutputStream os = new ByteArrayOutputStream();
    Main.solve(
        new ByteArrayInputStream(input.getBytes()), new PrintStream(os)
    );
    return os.toString().replace(System.lineSeparator(), "\n");
  }

}